Add 'has_occlusion' parameter with default value True
This commit introduces a new parameter 'has_occlusion' with a default value of True. The tracking mechanism now stops early under two conditions: when more than half of the tracking instances are lost, or when there is no occlusion detected in the video and at least one instance is lost during tracking. In other cases, the system will log a message indicating the event.