Add dialog to file menu for loading Sleap H5 files
This commit introduces a new feature to the file menu, allowing users to load Sleap H5 files. These files contain pose keypoints, and the selected Sleap H5 file can be converted into Labelme JSON format files. It is assumed that both the Sleap H5 file and the corresponding video file (in MP4 format) are located in the same folder.

The dialog provides a convenient way for users to select the desired Sleap H5 file for conversion, enhancing the usability of the application.